typescript(vx-demo): re-write Responsive + Lines demos in TypeScript
